Moved the event thread input the management of the Screen class.
